- Remove comments after discussion with Kyle san - Refactored some code
