+++ ui/src/main.tsx
@@ -1,7 +1,6 @@
 import { RouterProvider, createRouter } from "@tanstack/react-router";
 import { createRoot } from "react-dom/client";
 
-import { AuthProvider, AuthenticatedPage } from "./auth/index.tsx";
 import loadConfig from "./config";
 import { routeTree } from "./routeTree.gen.ts";
 
@@ -14,12 +13,5 @@ loadConfig().finally(async () => {
   if (!element) throw new Error("root element is not found");
 
   const root = createRoot(element);
-  root.render(
-    <AuthProvider>
-      {/* TODO: Not the correct way but works. Need to refactor this. */}
-      <AuthenticatedPage>
-        <RouterProvider router={router} />
-      </AuthenticatedPage>
-    </AuthProvider>,
-  );
+  root.render(<RouterProvider router={router} />);
 });
